+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix: drag-canvas behavior support scalable range, closes: #2136;
 - fix: TreeGraph changeData clear all states, closes: #2173;
 - chore: auto zoom tooltip & contextMenu component when zoom-canvas;
+- chore: upgrade @antv/g-canvas;
 
 #### 3.8.1
 - fix: update edge states with updateItem problem, closes: #2142;