+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - feat: destroyLayout API for graph, closes: #2140;
 - feat: clustering for force layout, closes: #2196;
 - fix: svg renderer minimap hidden elements probem, closes: #2174;
+- feat: add extra parameter graph for menu plugin, closes: #2204;
 
 #### 3.8.1
 - fix: update edge states with updateItem problem, closes: #2142;